- Centralizing the test data used in test_dataview, test_galleryview, test_columnview, test_itemsview, test_pickingviewdefault and test_pickingviewfilament
