+++ datavis/tests/__init__.py
@@ -1,2 +1,4 @@
 
-from . test_commons import TestBase, TestView, TestData
+from . test_commons import (TestBase, TestView, TestData, createSlicesModel,
+                            SimpleItemsModel, createTableModel,
+                            createPickerDataModel)
