+++ datavis/tests/test_slicesview.py
@@ -1,9 +1,6 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
 
-import numpy as np
-import pyqtgraph as pg
-
 import datavis as dv
 
 
@@ -14,30 +11,10 @@ class TestSlicesView(dv.tests.TestView):
         return ['']
 
     def createView(self):
-        return dv.views.SlicesView(None,
-                                   model=SimpleSlicesModel((512, 512), 50),
+        model = dv.tests.createSlicesModel((512, 512), 50)
+        return dv.views.SlicesView(None, model=model,
                                    text='Slice, image number: ',
                                    currentValue=1)
-
-
-class SimpleSlicesModel(dv.models.SlicesModel):
-    """ Example class implementing a simple slices model """
-
-    def __init__(self, imgSize, size):
-        """
-        Creates an StackModel.
-        :param imgSize: (tupple) The image size
-        :param size: (int) The number of images
-        """
-        dv.models.SlicesModel.__init__(self)
-        self._imgData = []
-        self._dim = imgSize[0], imgSize[1], size
-        self._imgData = [
-            pg.gaussianFilter(np.random.normal(size=imgSize),
-                              (5, 5)) for i in range(size)]
-
-    def getData(self, i=-1):
-        return None if i == -1 else self._imgData[i]
 
 
 if __name__ == '__main__':
